requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
certifi==2024.2.2
click==8.1.3
cloudinary==1.38.0
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
flake8==6.0.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Sijax==0.4.1
Flask-WTF==1.1.1
future==0.18.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
peewee==3.16.2
platformdirs==3.6.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyMySQL==1.0.3
python-decouple==3.8
pytz==2023.3
rich==13.4.2
shellingham==1.5.0.post1
Sijax==0.3.2
six==1.16.0
termcolor==2.3.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.2.0
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
